+++ proxy/process/browser.js
@@ -1,0 +1,5 @@
+import api from '../process.js';
+
+export default api;
+
+export * from '../process.js';